fix - killing enemyes
